Set user color in auth endpoint
