+++ apps/app/app/api/collaboration/auth/route.ts
@@ -1,5 +1,26 @@
 import { auth, currentUser } from '@clerk/nextjs/server';
 import { authenticate } from '@repo/collaboration/auth';
+import { tailwind } from '@repo/tailwind-config';
+
+const COLORS = [
+  tailwind.theme.colors.red[500],
+  tailwind.theme.colors.orange[500],
+  tailwind.theme.colors.amber[500],
+  tailwind.theme.colors.yellow[500],
+  tailwind.theme.colors.lime[500],
+  tailwind.theme.colors.green[500],
+  tailwind.theme.colors.emerald[500],
+  tailwind.theme.colors.teal[500],
+  tailwind.theme.colors.cyan[500],
+  tailwind.theme.colors.sky[500],
+  tailwind.theme.colors.blue[500],
+  tailwind.theme.colors.indigo[500],
+  tailwind.theme.colors.violet[500],
+  tailwind.theme.colors.purple[500],
+  tailwind.theme.colors.fuchsia[500],
+  tailwind.theme.colors.pink[500],
+  tailwind.theme.colors.rose[500],
+];
 
 export const POST = async () => {
   const user = await currentUser();
@@ -16,6 +37,7 @@ export const POST = async () => {
       name:
         user.fullName ?? user.emailAddresses.at(0)?.emailAddress ?? undefined,
       avatar: user.imageUrl ?? undefined,
+      color: COLORS[Math.floor(Math.random() * COLORS.length)],
     },
   });
 };
